WIP moved test cases to respective categories and wrote results to a simple text files. Left over cases are in tests/old/ and need to be addressed.
