+++ tests/mdao/mdao_integration_test_01.py
@@ -15,7 +15,8 @@ import unittest
 class MDAOTest(unittest.TestCase):
     def test_mdao(self):
 
-        print("mdao integration test")
+        # print("mdao integration test")
+        return
 
 if __name__ == '__main__':
     unittest.main()
