+++ tests/mission/mission_integration_test_03.py
@@ -1,19 +1,19 @@
 # Andy Torres
 # University of Central Florida
 # Department of Mechanical and Aerospace Engineering
-# Last Changed: 12-05-23
+# Last Changed: 03-16-24
 
 # ========================
-# PyRPOD: test/test_case_11.py
+# PyRPOD: tests/mission/mission_integration_test_03.py
 # ========================
-# Test case to contour the burn plot graph across various thrust and ISP values. (NEEDS TLC)
+# A brief test case to calculate RCS perfomance for a given flight plan approximating Δv requirements.
 
 import test_header
 import unittest, os, sys
 from pyrpod import LogisticsModule, MissionPlanner
 
-class BurnTimeContourChecks(unittest.TestCase):
-    def test_burn_time_contour_plots(self):
+class FlightPlanChecks(unittest.TestCase):
+    def test_rcs_flight_performance(self):
 
         # set case directory
         case_dir = '../case/flight_envelopes/'
@@ -30,16 +30,15 @@ class BurnTimeContourChecks(unittest.TestCase):
         # Load in thruster configuration data from text file
         lm.set_thruster_config()
 
-        # Draco/Hypergolic thrusters
+        # Assign properties of Draco/Hypergolic thrusters
         lm.add_thruster_performance(400, 300)
         lm.assign_thruster_groups()
 
-        # Read in flight data and plot burntime for a given Δv requirement.
-        # Graph is contoured according to various ISP values.
+        # Calculate simple 1D flight performance
         mp = MissionPlanner.MissionPlanner(case_dir)
         mp.set_lm(lm)
         mp.read_flight_plan()
-        # mp.plot_burn_time_contour(1194)
+        mp.calc_flight_performance()
 
 if __name__ == '__main__':
     unittest.main()
