+++ tests/mission/mission_integration_test_04.py
@@ -1,31 +1,19 @@
 # Andy Torres
 # University of Central Florida
 # Department of Mechanical and Aerospace Engineering
-# Last Changed: 12-05-23
+# Last Changed: 03-16-24
 
 # ========================
-# PyRPOD: test/test_case_06.py
+# PyRPOD: tests/mission/mission_integration_test_04.py
 # ========================
-# Test case to graph a thrust vs time or distance required given design requirements
-# and create a flight envelope to establish thrust requirements.
-
-# Given Reuirements
-# 1. Change in velocity (dV and /or dw)
-# 2. System mass properties
-# 3. Time or distance limits
-
-# Desired outputs
-# 1. Graph Thrust vs Time reuired.
-# 2. Graph Thrust vs Distance required.
-# 3. Use time and distance limits to create flight envelope data.
-# 4. Add data points for relevant thruster technologies.
+# Test case to analyze notional (1D transation + rotation) approach. (NEEDS TLC)
 
 import test_header
 import unittest, os, sys
 from pyrpod import LogisticsModule, MissionPlanner
 
-class ThrustEnvelopeChecks(unittest.TestCase):
-    def test_thrust_envelope_plot(self):
+class OneDimRotApproach(unittest.TestCase):
+    def test_1d_rot_approach_performance(self):
 
         # set case directory
         case_dir = '../case/flight_envelopes/'
@@ -46,11 +34,11 @@ class ThrustEnvelopeChecks(unittest.TestCase):
         lm.add_thruster_performance(400, 300)
         lm.assign_thruster_groups()
 
-        # Read in flight data and plot delta mass contoured for various Δv requirements.
+        # Calculate simple 1D flight performance
         mp = MissionPlanner.MissionPlanner(case_dir)
         mp.set_lm(lm)
         mp.read_flight_plan()
-        # mp.plot_thrust_envelope()
+        mp.calc_flight_performance()
         
 
 if __name__ == '__main__':
