+++ tests/mission/mission_integration_test_05.py
@@ -1,10 +1,10 @@
 # Andy Torres
 # University of Central Florida
 # Department of Mechanical and Aerospace Engineering
-# Last Changed: 12-05-23
+# Last Changed: 03-16-24
 
 # ========================
-# PyRPOD: test/test_case_07.py
+# PyRPOD: tests/mission/mission_integration_test_05.py
 # ========================
 # Test case to graph a thrust vs time or distance required given design requirements
 # and create a flight envelope to establish thrust requirements.
@@ -15,20 +15,17 @@
 # 3. Time or distance limits
 
 # Desired outputs
-# TODO: these are bad. need to think about them more.
-# 1. Graph ISP vs Fuel reuired.
+# 1. Graph Thrust vs Time reuired.
 # 2. Graph Thrust vs Distance required.
 # 3. Use time and distance limits to create flight envelope data.
 # 4. Add data points for relevant thruster technologies.
-# 5. Given a delta V and W requirement, comparison of thrust. isp, and mass flow rate. (M3) (Fuel Usage)
-
 
 import test_header
 import unittest, os, sys
 from pyrpod import LogisticsModule, MissionPlanner
 
-class DeltaMassChecks(unittest.TestCase):
-    def test_delta_m_plots(self):
+class ThrustEnvelopeChecks(unittest.TestCase):
+    def test_thrust_envelope_plot(self):
 
         # set case directory
         case_dir = '../case/flight_envelopes/'
@@ -49,11 +46,11 @@ class DeltaMassChecks(unittest.TestCase):
         lm.add_thruster_performance(400, 300)
         lm.assign_thruster_groups()
 
+        # Read in flight data and plot delta mass contoured for various Δv requirements.
         mp = MissionPlanner.MissionPlanner(case_dir)
         mp.set_lm(lm)
         mp.read_flight_plan()
-
-        # mp.plot_delta_m(1885)
+        # mp.plot_thrust_envelope()
         
 
 if __name__ == '__main__':
