+++ tests/mission/mission_integration_test_06.py
@@ -1,19 +1,33 @@
 # Andy Torres
 # University of Central Florida
 # Department of Mechanical and Aerospace Engineering
-# Last Changed: 12-05-23
+# Last Changed: 03-16-24
 
 # ========================
-# PyRPOD: test/test_case_05.py
+# PyRPOD: tests/mission/mission_integration_test_06.py
 # ========================
-# Test case to analyze notional (1D transation + rotation) approach. (NEEDS TLC)
+# Test case to graph a thrust vs time or distance required given design requirements
+# and create a flight envelope to establish thrust requirements.
+
+# Given Reuirements
+# 1. Change in velocity (dV and /or dw)
+# 2. System mass properties
+# 3. Time or distance limits
+
+# Desired outputs
+# TODO: these are bad. need to think about them more.
+# 1. Graph ISP vs Fuel reuired.
+# 2. Graph Thrust vs Distance required.
+# 3. Use time and distance limits to create flight envelope data.
+# 4. Add data points for relevant thruster technologies.
+# 5. Given a delta V and W requirement, comparison of thrust. isp, and mass flow rate. (M3) (Fuel Usage)
 
 import test_header
 import unittest, os, sys
 from pyrpod import LogisticsModule, MissionPlanner
 
-class OneDimRotApproach(unittest.TestCase):
-    def test_1d_rot_approach_performance(self):
+class DeltaMassChecks(unittest.TestCase):
+    def test_delta_m_plots(self):
 
         # set case directory
         case_dir = '../case/flight_envelopes/'
@@ -34,12 +48,11 @@ class OneDimRotApproach(unittest.TestCase):
         lm.add_thruster_performance(400, 300)
         lm.assign_thruster_groups()
 
-        # Calculate simple 1D flight performance
         mp = MissionPlanner.MissionPlanner(case_dir)
         mp.set_lm(lm)
         mp.read_flight_plan()
-        mp.calc_flight_performance()
-        
 
+        # mp.plot_delta_m(1885)
+        
 if __name__ == '__main__':
     unittest.main()
