+++ tests/mission/mission_integration_test_07.py
@@ -1,19 +1,19 @@
 # Andy Torres
 # University of Central Florida
 # Department of Mechanical and Aerospace Engineering
-# Last Changed: 12-05-23
+# Last Changed: 03-16-24
 
 # ========================
-# PyRPOD: test/test_case_04.py
+# PyRPOD: tests/mission/mission_integration_test_07.py
 # ========================
-# Test case to analyze notional 1DOF approach. (WIP)
+# Test case to contour the burn plot graph across various thrust and ISP values. (NEEDS TLC)
 
 import test_header
 import unittest, os, sys
 from pyrpod import LogisticsModule, MissionPlanner
 
-class OneDimTransApproachChecks(unittest.TestCase):
-    def test_1d_approach_performance(self):
+class BurnTimeContourChecks(unittest.TestCase):
+    def test_burn_time_contour_plots(self):
 
         # set case directory
         case_dir = '../case/flight_envelopes/'
@@ -34,10 +34,12 @@ class OneDimTransApproachChecks(unittest.TestCase):
         lm.add_thruster_performance(400, 300)
         lm.assign_thruster_groups()
 
+        # Read in flight data and plot burntime for a given Δv requirement.
+        # Graph is contoured according to various ISP values.
         mp = MissionPlanner.MissionPlanner(case_dir)
         mp.set_lm(lm)
         mp.read_flight_plan()
-        
+        # mp.plot_burn_time_contour(1194)
 
 if __name__ == '__main__':
     unittest.main()
