+++ tests/rpod/rpod_integration_test_01.py
@@ -1,21 +1,43 @@
 # Andy Torres
 # University of Central Florida
 # Department of Mechanical and Aerospace Engineering
-# Last Changed: 12-05-23
-
+# Last Changed: 03-16-24
 
 # ========================
-# PyRPOD: test/test_case_01.py
+# PyRPOD: tests/rpod/rpod_integration_test_01.py
 # ========================
-# A brief test case to calculate the 6DOF performance of each individual thruster in the LM
+# Test case to analyze notional 1DOF approach. (WIP)
 
 import test_header
-import unittest
+import unittest, os, sys
+from pyrpod import LogisticsModule, MissionPlanner
 
-class MDAOTest(unittest.TestCase):
-    def test_mdao(self):
+class OneDimTransApproachChecks(unittest.TestCase):
+    def test_1d_approach_performance(self):
 
-        print("rpod integration test")
+        # set case directory
+        case_dir = '../case/flight_envelopes/'
+
+        # Instantiate LogisticModule object.
+        lm = LogisticsModule.LogisticsModule(case_dir)
+
+        # Define LM mass distrubtion properties.
+        m = 0.45*30000 # lb converted to kg
+        h = 14 # m
+        r = 4.0/2.0 # m
+        lm.set_inertial_props(m, h, r)
+
+        # Load in thruster configuration data from text file
+        lm.set_thruster_config()
+
+        # Draco/Hypergolic thrusters
+        lm.add_thruster_performance(400, 300)
+        lm.assign_thruster_groups()
+
+        mp = MissionPlanner.MissionPlanner(case_dir)
+        mp.set_lm(lm)
+        mp.read_flight_plan()
+        
 
 if __name__ == '__main__':
     unittest.main()
