+++ tests/rpod/rpod_integration_test_03.py
@@ -1,19 +1,20 @@
 # Andy Torres
 # University of Central Florida
 # Department of Mechanical and Aerospace Engineering
-# Last Changed: 12-05-23
+# Last Changed: 03-16-24
 
 # ========================
-# PyRPOD: test/test_case_12.py
+# PyRPOD: tests/rpod/rpod_integration_test_03.py
 # ========================
-# Test case to contour the propellant usage across the various Δv in a given flight plan.
+# Test case to analyze Keep Out Zone Impingement. (WIP)
+
 
 import test_header
 import unittest, os, sys
 from pyrpod import LogisticsModule, MissionPlanner
 
-class DeltaMassContourChecks(unittest.TestCase):
-    def test_delta_m_plots(self):
+class KeepOutZoneChecks(unittest.TestCase):
+    def test_keep_out_zone(self):
 
         # set case directory
         case_dir = '../case/flight_envelopes/'
@@ -27,15 +28,17 @@ class DeltaMassContourChecks(unittest.TestCase):
         r = 4.0/2.0 # m
         lm.set_inertial_props(m, h, r)
 
-        # Draco/Hypergolic thrusters
+        # Load in thruster configuration data from text file
+        lm.set_thruster_config()
+
+        # Draco/Hypergolic thrusters.
         lm.add_thruster_performance(400, 300)
         lm.assign_thruster_groups()
 
-        # Read in flight data and plot delta mass contoured for various Δv requirements.
         mp = MissionPlanner.MissionPlanner(case_dir)
         mp.set_lm(lm)
         mp.read_flight_plan()
-        # mp.plot_delta_m_contour()
+
 
 if __name__ == '__main__':
     unittest.main()
