+++ tests/rpod/rpod_unit_test_03.py
@@ -1,13 +1,12 @@
 # Andy Torres
 # University of Central Florida
 # Department of Mechanical and Aerospace Engineering
-# Last Changed: 12-14-23
+# Last Changed: 03-16-24
 
 # ========================
-# PyRPOD: test/test_case_13.py
+# PyRPOD: tests/rpod/rpod_unit_test_03.py
 # ========================
-# Test case for converting STL data to VTK data.
-# This is accomplished by checking for the proper data format of VTK files.
+# Test case for producing JFH data accoring to a supplied equation.
 
 import test_header
 import unittest, os, sys
