--- tests/rpod/rpod_verification_test_01.py
@@ -1,21 +1,0 @@
-# Andy Torres
-# University of Central Florida
-# Department of Mechanical and Aerospace Engineering
-# Last Changed: 12-05-23
-
-
-# ========================
-# PyRPOD: test/test_case_01.py
-# ========================
-# A brief test case to calculate the 6DOF performance of each individual thruster in the LM
-
-import test_header
-import unittest
-
-class MDAOTest(unittest.TestCase):
-    def test_mdao(self):
-
-        print("rpod verification test")
-
-if __name__ == '__main__':
-    unittest.main()
